Update the change password/pin scenes to UI2
Changing/setting passwords or pins can now be routed to single files
regardless of whether it's an exisiting account or a new account.

Updated `react-redux` version to ^7.2.4 in order for functional
components to consume the `useDispatch` hook.

Note that there will be a white space at the bottom of the change
password/pin scenes. To resolve that, we would need to change the
`background` prop to `theme` in the `SceneWrapper` for the respective
file in the GUI repo.