Decouple ModalFooter from ModalScrollView
This dependency doesn't make sense and is causing styling issues on a case-by-case basis.
ModalFooter is visually a part of the ThemedModal in general, and should be managed in that outermost component.

The gradient fadeout should be the only thing coupled to the ModalScrollArea, as the gradient's sole purpose is to signal to the user that there is more scroll-able content. By coupling the gradient fadeout to the ModalScrollArea, we also overcome the strange styling where the ModalScrollArea can bleed far into the close button.

Finally, fadeout should never be optional. We should always show a gradient for scrollable content that overflows the screen. If the content is not scrollable, we should never show a gradient. While this is a worthwhile change for any non-modal ScrollView - gui or loginui - this PR aims to limit further scope creep.