Change TextInputModal to use KeyboardAwareScrollView
Additionally, pagingEnabled seems to conflict with
KeyboardAwareScrollView, causing a large overscroll/underscroll when
focusing/defocusing the control in iOS.

Unaddressed issues remain regarding Android/iOS inconsistency. iOS
specifically will auto-scroll the field when focused, but has strange
behavior where if invalid input is entered, the scroll behaves as if the
keyboard defocuses, even though the keyboard remains.

This fix at least addresses the bug of the close button overlapping the
submit button in all cases, but a larger scale change is required to
fully fix all cases of styling for android, ios, OutlinedTextInputFocus,
and KeyboardAwareScrollView.