Fix button margins on ChangeRecoveryScene There's a bug in ButtonUi4 where the margins/padding are being applied twice.
This issue is not apparent when using ButtonsViewUi4 because no custom margins/padding is being applied in that case.

Since we're hacking buttons for display on this scene, and the fact that we incorrectly have two equal primary actions, just apply custom margins that make it work to avoid regressions for other callers.