Convert the ChallengeModal to a legacy Airship modal
We cannot use the new modal, because the react-native-gesture-handler and blur features are messing with the WebView on Android. The blur seems to be re-rendering each time the WebView contents change, which is making the puzzle piece slider super-janky. Finally, there seems to be a layout issue that is making the WebView content invisible to start (but only on Android). We already had to disable the modal-close gesture in an earlier PR.

Rather than deal with these problems, let's put it back the way it was.