Fix recovery date off by one day bug
The date picker would return the date picked at the current local time
based on the device's locale (timezone). Because `toISOString` was used,
the resulting date my be different then the picked date because the day
in UTC may be a different day then what the current device is in.

To solve this, use `getDate/Month/FullYear` methods which respect locale
to format the ISO formatted date string. This has been abstracted into
a utility that is only scoped to the DateModel component.